move CalibrateTask from ProcessImageTask into ProcessCcdTask
CalibrateTask seems to have been included in ProcessImageTask before the
days of CoaddPsf, when we needed to be able to generate a Psf for a coadd
using the coadd; this is no longer necessary, and its presence in
ProcessImageTask stands in the way of further refactoring to support the
plan for coadd multi-band deblending and measurement.

In order to implement matching between the detected sources and the
reference catalog, I've added an AstrometryTask to ProcessImageTask, since
it can no longer rely on the one in CalibrateTask (well, with try/except
it could, but that would be undue familiarity with a subclass).  This
change does change the config hierarchy, and so will break use of
processCcd on data repositories with existing data (if not using
"--clobber-config").